Add throw exceptions in methods

+++ src/main/java/se/iths/rest/StudentController.java
@@ -7,6 +7,7 @@ import se.iths.service.StudentService;
 
 import javax.ejb.DuplicateKeyException;
 import javax.inject.Inject;
+import javax.mail.MethodNotSupportedException;
 import javax.ws.rs.*;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
@@ -39,18 +40,6 @@ public class StudentController {
         throw new NotFoundException();
     }
 
-    private boolean findDuplicate(Long id) {
-        List<Student> existingStudents = studentService.getStudents();
-        boolean conflict = false;
-        for (Student s: existingStudents) {
-            if (s.getId().equals(id)) {
-                conflict = true;
-                break;
-            }
-        }
-        return conflict;
-    }
-
     @Path("{id}")
     @GET
     public Response getStudent(@PathParam("id") Long id) {
@@ -60,6 +49,7 @@ public class StudentController {
         return Response.ok(student)
                 .build();
     }
+
     @Path("")
     @GET
     public Response getStudents() {
@@ -67,10 +57,13 @@ public class StudentController {
         return Response.ok(students)
                 .build();
     }
+
     @Path("findbyname")
     @GET
     public Response getStudents(@QueryParam("last-name") String name) {
         List<Student> students = studentService.getStudents(name);
+        if (students.isEmpty())
+            throw new StudentNotFoundException();
         return Response.ok(students)
                 .build();
     }
@@ -83,14 +76,37 @@ public class StudentController {
                 .status(Response.Status.NO_CONTENT)
                 .build();
     }
+    @Path("")
+    @PATCH
+    public void illegalPathUpdate() throws MethodNotSupportedException {
+        throw new MethodNotSupportedException();
+    }
     @Path("{id}")
     @DELETE
     public Response removeStudent(@PathParam("id") Long id) {
+        if (!findDuplicate(id))
+            throw new StudentNotFoundException();
         studentService.removeStudent(id);
         return Response.ok()
                 .status(Response.Status.NO_CONTENT)
                 .expires(Date.from(Instant.now()))
                 .build();
+    }
+    @Path("")
+    @PATCH
+    public void illegalPathDelete() throws MethodNotSupportedException {
+        throw new MethodNotSupportedException();
+    }
+    private boolean findDuplicate(Long id) {
+        List<Student> existingStudents = studentService.getStudents();
+        boolean conflict = false;
+        for (Student s: existingStudents) {
+            if (s.getId().equals(id)) {
+                conflict = true;
+                break;
+            }
+        }
+        return conflict;
     }
 }
